feat: Missing links redirect to search results
Problem

When you go to a missing link it doesnt return an
error or actually redirect you.

Solution

When a missing link is hit, return a 404 page
with s reasonable search result (top 20 ordered
by edit distance) to get to the intended page.

Result

If you typo when going to a link you will be
able to find what you're looking for if the
correct link is similar to what you typed.